Move the LogCat integration to Sentry/Platofrms/Android
