+++ src/Sentry.Maui/Internal/SentryMauiOptionsSetup.cs
@@ -33,13 +33,6 @@ internal class SentryMauiOptionsSetup : IConfigureOptions<SentryMauiOptions>
         // We'll use an event processor to set things like SDK name
         options.AddEventProcessor(new SentryMauiEventProcessor(options));
 
-#if ANDROID
-        if(options.LogCatIntegration != LogCatIntegrationType.None)
-        {
-            options.AddEventProcessor(new SentryAndroidLogsEventProcessor(options.LogCatIntegration));
-        }
-#endif
-
 #if !PLATFORM_NEUTRAL
         // We can use MAUI's network connectivity information to inform the CachingTransport when we're offline.
         options.NetworkStatusListener = new MauiNetworkStatusListener(Connectivity.Current, options);
