+++ src/Sentry/Platforms/Android/BindableSentryOptions.cs
@@ -1,4 +1,6 @@
 // ReSharper disable once CheckNamespace
+using Sentry.Android;
+
 namespace Sentry;
 
 internal partial class BindableSentryOptions
@@ -33,6 +35,7 @@ internal partial class BindableSentryOptions
         public TimeSpan? ReadTimeout { get; set; }
         public bool? EnableAndroidSdkTracing { get; set; }
         public bool? EnableAndroidSdkBeforeSend { get; set; }
+        public LogCatIntegrationType? LogCatIntegration { get; set; }
 
         public void ApplyTo(SentryOptions.AndroidOptions options)
         {
@@ -59,6 +62,7 @@ internal partial class BindableSentryOptions
             options.ReadTimeout = ReadTimeout ?? options.ReadTimeout;
             options.EnableAndroidSdkTracing = EnableAndroidSdkTracing ?? options.EnableAndroidSdkTracing;
             options.EnableAndroidSdkBeforeSend = EnableAndroidSdkBeforeSend ?? options.EnableAndroidSdkBeforeSend;
+            options.LogCatIntegration = LogCatIntegration ?? options.LogCatIntegration;
         }
     }
 }
